sataserver: add sataserver support to the VMM
This commit combines a number of commits which do the following:
    * Add virtio blk support to vmm
    * Separate virtio net into its own file
    * Link satadrivers into vmm library so VMM can use virtio-blk
    * Enable size max feature to virtio blk requests
    * Fix data overflow in io port in
    * Refactor virtio blk code in libsel4vmm
    * Cast values to avoid compiler warnings
    * Add mutex to virtio blk requests

CCDC-GVSC DISTRIBUTION A.  Approved for public release; distribution
unlimited. OPSEC#4481.
